Move Logger to logging.py and add log readers
