+++ asr/__init__.py
@@ -3,5 +3,6 @@ from asr.readers import read_ris
 from asr.review import review, review_oracle, review_simulate
 from asr.utils import load_data, text_to_features
 from asr.models.embedding import load_embedding, sample_embedding
+from asr.logging import read_log, read_logs_from_dir
 
 __version__ = "0.0.1"
