+++ asr/utils.py
@@ -1,15 +1,11 @@
 # Cpython dependencies
-import json
+import os
+import shutil
+from configparser import ConfigParser
 from os import environ
 from pathlib import Path
-import shutil
-from datetime import datetime
-import os
-from configparser import ConfigParser
 
 # external dependencies
-import numpy as np
-
 import pandas as pd
 
 
@@ -112,129 +108,6 @@ def text_to_features(sequences, num_words=20000, max_sequence_length=1000,
     return x, word_index
 
 
-class Logger(object):
-    """Class for logging the Systematic Review"""
-
-    def __init__(self, X=None):
-        super(Logger, self).__init__()
-
-        self.X = X
-
-        # since python 3, this is an ordered dict
-        self._log_dict = {"time": {"start_time": str(datetime.now())}}
-
-    def __str__(self):
-
-        self._print_logs()
-
-    def _print_logs(self):
-        self._log_dict["time"]["end_time"] = str(datetime.now())
-        s = "Logs of the Systematic Review process:\n"
-        for i, value in self._log_dict.items():
-            s += f"Query {i} - Reduction {value}"
-
-        return s
-
-    def _add_log(self, new_dict, i):
-        # Find the first number that is not logged yet.
-        if i is None:
-            i = 0
-            while i in self._log_dict:
-                # If the keys of the new dictionary don't exist, this is it.
-                if set(new_dict.keys()).isdisjoint(self._log_dict[i].keys()):
-                    break
-                i += 1
-
-        if i not in self._log_dict:
-            self._log_dict[i] = {}
-
-        self._log_dict[i].update(new_dict)
-
-    def add_settings(self, settings):
-        self._log_dict["settings"] = settings
-
-    def add_training_log(self, indices, labels, i=None):
-        """Add training indices and their labels.
-
-        Arguments
-        ---------
-        indices: list, np.array
-            A list of indices used for training.
-        labels: list
-            A list of labels corresponding with the training indices.
-        i: int
-            The query number.
-        """
-
-        # ensure that variables are serializable
-        if isinstance(indices, np.ndarray):
-            indices = indices.tolist()
-        if isinstance(labels, np.ndarray):
-            labels = labels.tolist()
-        new_dict = {'labelled': list(zip(indices, labels))}
-        self._add_log(new_dict, i)
-
-    def add_pool_predict(self, indices, pred, i=None):
-        """Add inverse pool indices and their labels.
-
-        Arguments
-        ---------
-        indices: list, np.array
-            A list of indices used for unlabeled pool.
-        pred: np.array
-            A list of predictions for those samples.
-        i: int
-            The query number.
-        """
-
-        if isinstance(indices, np.ndarray):
-            indices = indices.tolist()
-        pred = np.reshape(pred, -1)
-        if isinstance(pred, np.ndarray):
-            pred = pred.tolist()
-        new_dict = {'predictions': list(zip(indices, pred))}
-        self._add_log(new_dict, i)
-
-    def add_proba(self, indices, pred_proba, logname="pool_proba", i=None):
-        """Add inverse pool indices and their labels.
-
-        Arguments
-        ---------
-        indices: list, np.array
-            A list of indices used for unlabeled pool.
-        pred: np.array
-            Array of prediction probabilities for unlabeled pool.
-        i: int
-            The query number.
-        """
-
-        if isinstance(indices, np.ndarray):
-            indices = indices.tolist()
-        pred_proba = pred_proba[:, 1]
-        if isinstance(pred_proba, np.ndarray):
-            pred_proba = pred_proba.tolist()
-        new_dict = {logname: list(zip(indices, pred_proba))}
-        self._add_log(new_dict, i)
-
-    def save(self, fp):
-        """Save logs to file.
-
-        Arguments
-        ---------
-        fp: str
-            The file path to export the results to.
-
-        """
-        self._log_dict["time"]["end_time"] = str(datetime.now())
-        fp = Path(fp)
-
-        if fp.is_file:
-            fp.parent.mkdir(parents=True, exist_ok=True)
-
-        with fp.open('w') as outfile:
-            json.dump(self._log_dict, outfile, indent=2)
-
-
 def get_data_home(data_home=None):
     """Return the path of the ASR data dir.
 
@@ -296,4 +169,3 @@ def config_from_file(config_file):
             print(f"Warning: section [{sect}] is ignored in"
                   f" config file {config_file}")
     return settings
-
